move static check to refreshTokenHandler
